pyproject.toml
[project]
name = "data-engineering-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "airflow-operators>=0.11.0",
    "apache-airflow>=3.1.8",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "requests>=2.33.0",
]
